backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.0
requests==2.32.3
fastapi-cors==0.0.6
pymongo==4.10.1
httpx==0.28.1
pytest==8.3.4
pytest-cov==6.0.0
mcp>=1.6.0
pydantic>=2.7.2
openseespy>=3.7.0
numpy>=1.24.0
sectionproperties>=3.1.0
websockets>=12.0
pydantic[email]
passlib
python-jose